Update edit_node.  Fixes #14
Updates the edit_node method to accept all parameters accepted by the API.  Despite being a PATCH endpoint this still requires you to pass all fields in order to modify any of them.  Will create a common function that handles populating defaults from the API and update this to use it in the future.

Removes some Python 2 compatibility.