Add includes and params to client API.
Since the client API is limited in scope to an individual server there were only a few places it seemed prudent to add includes and params.  It's quite possible that some endpoints were overlooked, but I believe I got all the endpoints that cannot be updated in a way that is backwards compatible.